updated index parameters to size_t, overrided [] operator for const instances, updated Matrix constructor to handle ill-formed matrices, unit tests

+++ include/canvas.h
@@ -12,11 +12,11 @@ class Canvas
     private:
         std::vector<std::vector<Color>> _pixels;
     public:
-        Canvas(unsigned w, unsigned h): width(w), height(h){
+        Canvas(size_t w, size_t h): width(w), height(h){
             _pixels.resize(height, std::vector<Color>(width, Color(0, 0, 0)));
         };
-        unsigned width;
-        unsigned height;
+        size_t width;
+        size_t height;
         std::vector<Color>& operator[](const size_t i);
         string to_ppm(string filename);
 };
